Allow up to 16 parallel I/Os at a time
We open the same file 16 files to create 16 file descriptors each
with an independently-seekable offset. This allows up to 16 read
and write requests to be issued in parallel.

The aim is to emulate POSIX aio without actually using POSIX aio
(since the POSIX aio C interface is ugly and hard to bind)

Signed-off-by: David Scott <dave.scott@eu.citrix.com>