Add discard support
Discard (also TRIM also UNMAP) requests that a block device throw away
data in a range of sectors. This is used as an optimisations by SSDs
and by filesystems which support sparse files. It is guaranteed that
reads will return zeroes after a discard (also known as "Read Zero After
TRIM")

Signed-off-by: David Scott <dave@recoil.org>